model_workers/worker_framework/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "worker_framework"
version = "0.0.4"
authors = [
  { name="Yung-Hsiang Hu", email="iftnt1999@gmail.com" },
]
description = "A framework to serve and compose LLMs."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies=[
  'uvicorn==0.23.2',
  'starlette==0.31.1',
  'sse-starlette==1.6.5',
  'python-multipart==0.0.6',
  'pyyaml==6.0.1',
  'requests==2.31.0',
  'dacite==1.8.1',
  'prometheus-client==0.17.1'
]

[project.urls]
"Homepage" = "https://www.taide.tw/"
"Bug Tracker" = ""

[project.scripts]
worker-server = "worker_framework.main:main"